Added file widget and presentation
In this widget all that's left is creating the "compiler" in order to
write to it.
Something has to be done to the search box but it's not going to be in
the first release by the looks of it.
Then it's copy and paste for the style management widget.