Added pandoc functionality
Because I came w/o a plan I paid the price.
I currently don't know the scope and gravity of all the changes in this
commit.
Those who are known are
* Introduced Title and Icon to several settings widgets
* Added pandoc enabling to the behaviour widget
* Added the basic pandoc functionality to CentralWidget
* Added an htmlizer in Files edit for pandoc purposes

Because of the lack of planning I have made a mess and had several
different ideas that may or may not have left some code all over the
place. Thats gonna be a problem or maybe not.

Also one of the plans included usin the qtwebkit. That lead nowhere. You
should really look into a self hosting website for what you want. Which
you must admit it's kinda ironic given your hate for atom applications.
Fuck css support in qt. Hopefully on some future release it's fully
supported.

The lesson of the day is commit more. Also

DON'T DO SHIT WITHOUT A PLAN!!!

wingin it it's just more work in the end